+++ app/routes/components/Select/changelog.codegen.ts
@@ -9,4 +9,5 @@ export const changelog = ''
   + '- 2025-10-03: Fix missing dependencies in useEffect hooks, etc.\n'
   + '- 2025-10-11: Support `id` prop for better form integration. Make it be able to\n'
   + '  uncontrolled. Support `defaultValue` prop. Code formatting improvements.\n'
+  + '- 2025-10-26: Update styles (padding) to make it look better.\n'
   ;